move li.fun.generate_lr_geneset to rs
